+++ src/TeaPieLanguageServer.ts
@@ -2,13 +2,13 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
 
-import { XmlDocMember, XmlDocParser } from './XmlDocParser';
+import { XmlDocMember, XmlDocParser } from './modules/XmlDocParser';
 
 // Create output channel for logging
 let outputChannel: vscode.OutputChannel;
 
 // Remove this import if not used
-// import { fileURLToPath } from 'url';
+// import { fileURLToPath } from 'url';T
 
 export class TeaPieLanguageServer {
     private static instance: TeaPieLanguageServer;
@@ -105,7 +105,7 @@ export class TeaPieLanguageServer {
                                 const beforeDot = linePrefix.slice(0, -1).trim();
                                 outputChannel.appendLine(`[TeaPie] Providing completions for context: "${beforeDot}"`);
                                 const filteredMembers = XmlDocParser.getCompletionItems(this.xmlDocs, beforeDot);
-                                completionItems = filteredMembers.map(member => this.createCompletionItem(member));
+                                completionItems = filteredMembers.map((member: XmlDocMember) => this.createCompletionItem(member));
                             }
 
                             if (completionItems.length === 0) {
@@ -214,7 +214,7 @@ export class TeaPieLanguageServer {
                 const teaPieMembers = XmlDocParser.getMembersByType(this.xmlDocs, 'TeaPie.TeaPie');
                 outputChannel.appendLine(`[TeaPie] Found ${teaPieMembers.length} TeaPie.TeaPie members`);
                 if (teaPieMembers.length > 0) {
-                    outputChannel.appendLine('[TeaPie] Sample TeaPie members: ' + JSON.stringify(teaPieMembers.slice(0, 3).map(m => m.name)));
+                    outputChannel.appendLine('[TeaPie] Sample TeaPie members: ' + JSON.stringify(teaPieMembers.slice(0, 3).map((m: XmlDocMember) => m.name)));
                 }
             }
         } catch (error) {
