--- src/utils/variablesParser.ts
@@ -1,8 +1,0 @@
-export function parseVariablesFile(content: string): { [key: string]: any } {
-    try {
-        return JSON.parse(content);
-    } catch (error) {
-        console.error('Error parsing variables file:', error);
-        return {};
-    }
-}
